+++ src/routes/+page.svelte
@@ -47,7 +47,7 @@
 
 <a href="/test">Test page</a> | <a href="/crud">CRUD</a> |
 <a href="/super-debug">SuperDebug</a>
-| <a href="/nested">Nested</a>
+| <a href="/nested">Nested</a> | <a href="/multiple">Multiple</a>
 
 {#if $message}
   <h4 class:error={$page.status >= 400} class="message">{$message}</h4>